# server/docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Water tracker app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of water tracker app"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://projectwatertrackerteam4.onrender.com"
    }
  ],
...
  "paths": {
...
    "/water/day/{date}": {
      "get": {
        "tags": [
          "Water"
        ],
        "summary": "Get water records by day",
        "description": "This endpoint retrieves water records for a given date.",
        "operationId": "getWaterByDate",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "description": "The date for which to retrieve the water records (ISO format YYYY-MM-DD).",
            "required": true,
            "schema": {
              "type": "string",
              "example": "2024-12-16"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the water records for the given date.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found water records by this date."
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "waterRecords",
                        "recordsCount",
                        "totalDayWater",
                        "percentage"
                      ],
                      "properties": {
                        "waterRecords": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "_id",
                              "userId",
                              "date",
                              "amount",
                              "currentDailyNorm"
                            ],
                            "properties": {
                              "_id": {
                                "type": "string",
                                "description": "Unique identifier for the water record.",
                                "example": "64f5b992c4b3f1567e63a179"
                              },
                              "userId": {
                                "type": "string",
                                "description": "ID of the user who logged the water record.",
                                "example": "64f5b992c4b3f1567e63a178"
                              },
                              "date": {
                                "type": "string",
                                "format": "date-time",
                                "description": "The date and time of the water consumption record.",
                                "example": "2024-12-16T08:30:00Z"
                              },
                              "amount": {
                                "type": "number",
                                "description": "The amount of water consumed (in milliliters).",
                                "example": 300
                              },
                              "currentDailyNorm": {
                                "type": "number",
                                "description": "The current daily water consumption norm for the user (in milliliters).",
                                "example": 2000
                              }
                            }
                          }
                        },
                        "recordsCount": {
                          "type": "integer",
                          "description": "The number of records found for the given date.",
                          "example": 3
                        },
                        "totalDayWater": {
                          "type": "number",
                          "description": "The total amount of water consumed for the given date (in milliliters).",
                          "example": 900
                        },
                        "percentage": {
                          "type": "integer",
                          "description": "The percentage of the daily water consumption norm met for the given date.",
                          "example": 45
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "No records found for the given date.",
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
      "400": {
        "description": "Bad Request - Validation Error or Incorrect Input",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Validation Error! Invalid request body or parameters."
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid input data"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "NotFoundError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "The requested resource was not found!"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "error"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "InternalServerError"
                },
                "error": {
                  "type": "string",
                  "description": "Detailed error message for debugging purposes.",
                  "example": "Unexpected database error occurred"
                }
              }
            }
          }
        }
      }
    }
  }
}